requirements.test.txt
# Strictly for tests
pytest
pytest-cov<3.0.0
pytest-homeassistant-custom-component>=0.4.8
# From our manifest.json for our custom component
